src/unitTestExcercise/494findTargetSumWays_test.go
package unitTestExcercise

import "testing"

func Test_findTargetSumWays(t *testing.T) {
	type args struct {
		nums []int
		S    int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// TODO: Add test cases.
		{"1", args{[]int{1, 1, 1, 1, 1}, 3}, 5},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := findTargetSumWays(tt.args.nums, tt.args.S); got != tt.want {
				t.Errorf("findTargetSumWays() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_helper494(t *testing.T) {
	type args struct {
		nums []int
		S    int
		re   []byte
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// TODO: Add test cases.
		{"1", args{[]int{1, 1, 1, 1, 1}, 3, []byte{}}, 5},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := helper494(tt.args.nums, tt.args.S, tt.args.re); got != tt.want {
				t.Errorf("helper494() = %v, want %v", got, tt.want)
			}
		})
	}
}
